fix error message; improve tests
